Update Adafruit_BNO08x_RVC.cpp enabling grabbing most recent data from serial buffer.
My tests show that the BNO085 spits out data every 11ms when using the Feather ESP32 V2. If not polling the BNO085 less than approximately 18ms, the serial buffer will accumulate data from the BNO085. If the user is polling every 100ms for example, there will be at least 9 data packets in the buffer, and the latest read will get the data from 9 packets ago, not the most recent data.

It then takes multiple reads before being caught up to the most recent data, which if not handled, the user may never be retrieving the most recent data.

Even more importantly, because the buffer has had time to fill, it will have more than one packet of data. The current code only checks to see that there's at least 19 bytes in the buffer, but never checks to see if there's more than 19bytes.  The update proposed here handles that. It will go through all the fully formed packets in the buffer, discarding all but the most recent data packet.